requirements.txt
bencode.py
isbnlib
Pillow
tqdm
zstandard
